frontend/A32: Correct more IT state

+++ src/frontend/A32/translate/conditional_state.cpp
@@ -41,7 +41,7 @@ bool IsConditionPassed(TranslatorVisitor& v, IR::Cond cond) {
             v.cond_state = ConditionalState::Trailing;
         } else {
             if (cond == v.ir.block.GetCondition()) {
-                v.ir.block.SetConditionFailedLocation(v.ir.current_location.AdvancePC(v.current_instruction_size).AdvanceIT());
+                v.ir.block.SetConditionFailedLocation(v.ir.current_location.AdvancePC(static_cast<int>(v.current_instruction_size)).AdvanceIT());
                 v.ir.block.ConditionFailedCycleCount()++;
                 return true;
             }
@@ -72,7 +72,7 @@ bool IsConditionPassed(TranslatorVisitor& v, IR::Cond cond) {
 
     v.cond_state = ConditionalState::Translating;
     v.ir.block.SetCondition(cond);
-    v.ir.block.SetConditionFailedLocation(v.ir.current_location.AdvancePC(v.current_instruction_size).AdvanceIT());
+    v.ir.block.SetConditionFailedLocation(v.ir.current_location.AdvancePC(static_cast<int>(v.current_instruction_size)).AdvanceIT());
     v.ir.block.ConditionFailedCycleCount() = v.ir.block.CycleCount() + 1;
     return true;
 }
